♻️ engineering_team_runtime → agents 이동 + discord compat shim
변경 이유:
discord/engineering_team_runtime 는 discord transport 가 전혀 없는
engineering-agent deliberation/research 오케스트레이션 로직인데
discord/ 아래 잘못 위치해 agents↔discord 인위적 순환을 만든다.
agents/ 로 옮겨 순환의 한 축을 제거한다.

주요 변경 사항:
- src/yule_orchestrator/discord/engineering_team_runtime → agents/ 로 relocate.
- 모듈 내부 cross-package import 를 절대경로로 재작성:
  from ...agents → from yule_orchestrator.agents,
  from ...memory → from yule_orchestrator.memory.
- 기존 discord 경로에는 compat shim 패키지 (__init__ public 재노출 +
  _legacy sys.modules alias) 를 남겨 비-agents import 사이트 (discord/cli/tests)
  를 무손상 유지.

비고:
shim 으로 discord→agents (forward/legal 방향) 만 남는다.
agents-side importer 재배선은 후속 커밋에서 수행.
